requirements.txt
click==6.7
dlib==19.9.0
face-recognition==1.2.1
face-recognition-models==0.3.0
numpy==1.14.1
opencv-python==3.4.0.12
paho-mqtt==1.3.1
Pillow==5.0.0
scipy==1.0.0
